refactor: extract Toggle component to make it reusable
